# candidate-match-export/candidate-match/openapi.yaml
openapi: 3.0.3
info:
  title: Cloudberries Candidate Match API
  version: 1.1.0
  description: |
    API for candidate matching, CV retrieval, embeddings, and health checks.
    Includes normalized health responses and skill-based matching endpoint.
    Generated from controller sources as of this commit.
  contact:
    name: Cloudberries
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /chatbot/analyze:
    post:
      tags: [ Chat ]
      summary: Analyze content with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIAnalysisRequest'
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIResponseModel'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /chatbot/search:
    post:
      tags: [ Chat ]
      summary: AI-powered consultant search
      description: Search consultants using natural language with intelligent routing to STRUCTURED, SEMANTIC, HYBRID, or RAG search modes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatSearchRequest'
            examples:
              structured:
                summary: Structured search example
                value:
                  text: "Find consultants who know Kotlin and Spring"
                  topK: 10
              semantic:
                summary: Semantic search example  
                value:
                  text: "Experienced fullstack developer who can mentor juniors"
                  topK: 5
              forced_mode:
                summary: Forced mode example
                value:
                  text: "Java developer"
                  forceMode: "SEMANTIC"
                  topK: 3
              rag:
                summary: RAG search example
                value:
                  text: "Tell me about Thomas Andersen's experience with React"
                  topK: 5
              targeted:
                summary: Target a specific consultant and CV
                value:
                  consultantId: "thomas.andersen"
                  cvId: "default"
                  text: "Summarize key React experience and recent roles"
                  forceMode: "RAG"
                  topK: 3
      responses:
        '200':
          description: Search results with AI interpretation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSearchResponse'
        '400':
          description: Invalid request parameters
        default:
          $ref: '#/components/responses/ErrorResponse'
  /consultants:
    get:
      tags: [ Consultants ]
      summary: List consultants (paged)
      parameters:
        - in: query
          name: name
          schema:
            type: string
          description: Optional name filter (contains)
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paged consultants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageConsultantSummaryDto'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /consultants/with-cv:
    get:
      tags: [ Consultants ]
      summary: List all consultants with their CVs (normalized entities)
      parameters:
        - in: query
          name: onlyActiveCv
          description: If true, include only active CVs for each consultant
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Consultants with nested CV structures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsultantWithCvDto'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /consultants/with-cv/paged:
    get:
      tags: [ Consultants ]
      summary: List consultants with CVs (paged)
      parameters:
        - in: query
          name: onlyActiveCv
          schema:
            type: boolean
            default: false
          description: If true, include only active CVs for each consultant
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paged consultants with CVs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageConsultantWithCvDto'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /consultants/{userId}/cvs:
    get:
      tags: [ Consultants ]
      summary: List CVs for a consultant by userId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: onlyActiveCv
          required: false
          schema:
            type: boolean
            default: false
          description: If true, include only the active CV
      responses:
        '200':
          description: Consultant CVs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsultantCvDto'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /consultants/search:
    post:
      tags: [ Consultants ]
      summary: Relational search for consultants
      operationId: searchConsultantsRelational
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationalSearchRequest'
            example:
              name: "kotlin"
              skillsAll: [ "KOTLIN" ]
              skillsAny: [ "BACKEND" ]
              minQualityScore: 70
              onlyActiveCv: true
              pagination:
                page: 0
                size: 10
                sort: [ "name,asc" ]
      responses:
        '200':
          description: Paged consultants matching filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageConsultantWithCvDto'
        '400':
          description: Invalid search criteria
        default:
          $ref: '#/components/responses/ErrorResponse'
  /consultants/search/semantic:
    post:
      tags: [ Consultants ]
      summary: Semantic (vector) search for consultants
      operationId: searchConsultantsSemantic
      description: Embeds the provided text using the configured EmbeddingProvider and performs pgvector similarity search.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticSearchRequest'
            example:
              text: "Senior Kotlin developer with Spring experience"
              provider: "GOOGLE_GEMINI"
              model: "text-embedding-004"
              topK: 5
              minQualityScore: 80
              onlyActiveCv: true
              pagination:
                page: 0
                size: 10
                sort: [ "name,asc" ]
      responses:
        '200':
          description: Top-K semantically similar consultants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageConsultantWithCvDto'
        '400':
          description: Invalid search criteria
        '503':
          description: Embedding provider disabled or not configured
        default:
          $ref: '#/components/responses/ErrorResponse'
  /consultants/search/embedding-info:
    get:
      tags: [ Consultants ]
      summary: Get embedding provider information
      description: Returns information about the available embedding provider for semantic search
      responses:
        '200':
          description: Embedding provider information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbeddingProviderInfo'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  parameters:
    PageParam:
      in: query
      name: page
      description: Page number (0-indexed)
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      in: query
      name: size
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortParam:
      in: query
      name: sort
      description: Sort field(s), e.g. `name,asc`. Repeat for multi-sort.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    RelationalSearchRequest:
      type: object
      properties:
        name:
          type: string
          description: Name filter (contains match)
        skillsAll:
          type: array
          items:
            type: string
          description: Skills that must ALL be present (AND condition)
        skillsAny:
          type: array
          items:
            type: string
          description: Skills where ANY can be present (OR condition)
        minQualityScore:
          type: integer
          minimum: 0
          maximum: 100
          description: "Minimum CV quality score (0–100). Candidates without a score are treated as 0 for filtering."
        onlyActiveCv:
          type: boolean
          default: false
          description: Only include active CVs
        pagination:
          $ref: '#/components/schemas/PaginationDto'
    SemanticSearchRequest:
      type: object
      required: [ text ]
      properties:
        text:
          type: string
          description: Natural language search text
        provider:
          type: string
          default: "GOOGLE_GEMINI"
          description: Embedding provider (must match server configuration)
        model:
          type: string
          default: "text-embedding-004"
          description: Embedding model (must match server configuration)
        topK:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Maximum number of results to return
        minQualityScore:
          type: integer
          minimum: 0
          maximum: 100
          description: "Minimum CV quality score (0–100). Candidates without a score are treated as 0 for filtering."
        onlyActiveCv:
          type: boolean
          default: false
          description: Only include active CVs
        pagination:
          $ref: '#/components/schemas/PaginationDto'
    EmbeddingProviderInfo:
      type: object
      required: [ enabled, provider, model, dimension ]
      properties:
        enabled:
          type: boolean
          description: Whether semantic search is available
        provider:
          type: string
          description: Name of the embedding provider
        model:
          type: string
          description: Name of the embedding model
        dimension:
          type: integer
          description: Vector dimension of the embeddings
...
    AIAnalysisRequest:
      type: object
      required: [ content ]
      properties:
        content:
          type: string
          description: Content to analyze
    AIResponseModel:
      type: object
      properties:
        content:
          type: string
        modelUsed:
          type: string
    ConsultantSummaryDto:
      type: object
      required: [ userId, name, email, bornYear, defaultCvId ]
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        bornYear:
          type: integer
        defaultCvId:
          type: string
    PageConsultantSummaryDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ConsultantSummaryDto'
        number:
          type: integer
          description: Page number (0-indexed)
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        sort:
          type: object
          additionalProperties: true
        pageable:
          type: object
          additionalProperties: true
    ConsultantWithCvDto:
      type: object
      required: [ userId, name, cvId, skills, cvs ]
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        userId:
          type: string
        name:
          type: string
        cvId:
          type: string
        skills:
          type: array
          description: Top 3 skills for the consultant when available
          maxItems: 3
          items:
            type: string
        cvs:
          type: array
          items:
            $ref: '#/components/schemas/ConsultantCvDto'
    ConsultantCvDto:
      type: object
      required: [ active ]
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        versionTag:
          type: string
          nullable: true
        qualityScore:
          type: integer
          nullable: true
          minimum: 0
          maximum: 100
          description: "CV quality score (0–100); populated from cv_score.score_percent. May be null when no score exists."
        active:
          type: boolean
        keyQualifications:
          type: array
          items: { $ref: '#/components/schemas/KeyQualificationDto' }
        education:
          type: array
          items: { $ref: '#/components/schemas/EducationDto' }
        workExperience:
          type: array
          items: { $ref: '#/components/schemas/WorkExperienceDto' }
        projectExperience:
          type: array
          items: { $ref: '#/components/schemas/ProjectExperienceDto' }
        certifications:
          type: array
          items: { $ref: '#/components/schemas/CertificationDto' }
        courses:
          type: array
          items: { $ref: '#/components/schemas/CourseDto' }
        languages:
          type: array
          items: { $ref: '#/components/schemas/LanguageDto' }
        skillCategories:
          type: array
          items: { $ref: '#/components/schemas/SkillCategoryDto' }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/AttachmentDto' }
        industries:
          type: array
          items:
            type: string
    KeyQualificationDto:
      type: object
      properties:
        label: { type: string, nullable: true }
        description: { type: string, nullable: true }
    EducationDto:
      type: object
      properties:
        degree: { type: string, nullable: true }
        school: { type: string, nullable: true }
        fromYearMonth: { type: string, nullable: true }
        toYearMonth: { type: string, nullable: true }
    WorkExperienceDto:
      type: object
      properties:
        employer: { type: string, nullable: true }
        fromYearMonth: { type: string, nullable: true }
        toYearMonth: { type: string, nullable: true }
    ProjectExperienceDto:
      type: object
      properties:
        customer: { type: string, nullable: true }
        description: { type: string, nullable: true }
        longDescription: { type: string, nullable: true }
        fromYearMonth: { type: string, nullable: true }
        toYearMonth: { type: string, nullable: true }
        roles:
          type: array
          items: { $ref: '#/components/schemas/ProjectRoleDto' }
        skills:
          type: array
          items: { type: string }
    ProjectRoleDto:
      type: object
      properties:
        name: { type: string, nullable: true }
        description: { type: string, nullable: true }
    CertificationDto:
      type: object
      properties:
        name: { type: string, nullable: true }
        year: { type: integer, nullable: true }
    CourseDto:
      type: object
      properties:
        name: { type: string, nullable: true }
        organizer: { type: string, nullable: true }
        year: { type: integer, nullable: true }
    LanguageDto:
      type: object
      properties:
        name: { type: string, nullable: true }
        level: { type: string, nullable: true }
    SkillCategoryDto:
      type: object
      properties:
        name: { type: string, nullable: true }
        skills:
          type: array
          items: { $ref: '#/components/schemas/SkillInCategoryDto' }
    SkillInCategoryDto:
      type: object
      properties:
        name: { type: string, nullable: true }
        durationYears: { type: integer, nullable: true }
    AttachmentDto:
      type: object
      properties:
        fileName: { type: string, nullable: true }
        url: { type: string, nullable: true }
    PageConsultantWithCvDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ConsultantWithCvDto'
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        sort:
          type: object
          additionalProperties: true
        pageable:
          type: object
          additionalProperties: true
...
    PaginationDto:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
          default: 0
          description: Page number (0-indexed)
        size:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Number of items per page
        sort:
          type: array
          items:
            type: string
          description: Sort specifications (e.g. "name,asc", "id,desc")
    Problem:
      type: object
      description: Generic error payload
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # AI Chat Search Schemas
    ChatSearchRequest:
      type: object
      required:
        - text
      properties:
        conversationId:
          type: string
          description: Optional conversation ID to maintain context
          example: "conv-123"
        consultantId:
          type: string
          description: Optional consultant userId to target the query (used to bias or provide RAG context)
          example: "thomas.andersen"
        cvId:
          type: string
          description: Optional CV/resume id for the selected consultant; used to include CV JSON as context to the AI
          example: "default"
        text:
          type: string
          description: Natural language search text
          example: "Find consultants who know Kotlin and Spring"
        forceMode:
          type: string
          enum: ["STRUCTURED", "SEMANTIC", "HYBRID", "RAG"]
          description: Force a specific search mode
          example: "STRUCTURED"
        topK:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
          description: Maximum number of results to return
          example: 10
    ChatSearchResponse:
      type: object
      required:
        - mode
        - latencyMs
      properties:
        mode:
          type: string
          enum: ["STRUCTURED", "SEMANTIC", "HYBRID", "RAG"]
          description: Search mode used
          example: "STRUCTURED"
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
          description: Search results (for structured/semantic/hybrid modes)
        answer:
          type: string
          description: Generated answer text (for RAG mode)
          example: "Based on the consultant's CV..."
        sources:
          type: array
          items:
            $ref: '#/components/schemas/RAGSource'
          description: Sources used for RAG answers
        latencyMs:
          type: integer
          description: Response time in milliseconds
          example: 1250
        debug:
          $ref: '#/components/schemas/DebugInfo'
          description: Optional debug information
        conversationId:
          type: string
          description: Conversation ID for follow-up queries
        scoring:
          $ref: '#/components/schemas/ScoringInfo'
    SearchResult:
      type: object
      required:
        - consultantId
        - name
        - score
      properties:
        consultantId:
          type: string
          description: Consultant userId
        name:
          type: string
          description: Consultant name
          example: "Thomas Andersen"
        score:
          type: number
          format: double
          minimum: 0.0
          maximum: 1.0
          description: Relevance score (0-1)
          example: 0.87
        highlights:
          type: array
          items:
            type: string
          description: Text highlights from matching
        meta:
          type: object
          additionalProperties: true
          description: Additional metadata
    RAGSource:
      type: object
      required:
        - consultantId
        - consultantName
        - chunkId
        - text
        - score
      properties:
        consultantId:
          type: string
          description: Consultant userId
        consultantName:
          type: string
          description: Consultant name
        chunkId:
          type: string
          description: Chunk identifier
        text:
          type: string
          description: Source text excerpt
        score:
          type: number
          format: double
          description: Relevance score
          example: 0.92
        location:
          type: string
          description: CV section location
          example: "Experience"
    DebugInfo:
      type: object
      properties:
        interpretation:
          $ref: '#/components/schemas/QueryInterpretation'
          description: Query interpretation details
        timings:
          type: object
          additionalProperties:
            type: integer
          description: Timing breakdown
        extra:
          type: object
          additionalProperties: true
          description: Additional debug data
    ScoringInfo:
      type: object
      description: HYBRID scoring weights and formula
      properties:
        semanticWeight:
          type: number
          format: double
          example: 0.7
        qualityWeight:
          type: number
          format: double
          example: 0.3
        formula:
          type: string
          example: "combined = semanticWeight * semanticScore + qualityWeight * qualityScore"
...